requirements.txt
python-whois
colorama
subprocess